+++ lib/src/backend/js/backend_manager.dart
@@ -1,14 +1,14 @@
 import 'dart:html';
 import 'dart:indexed_db';
 
+import 'package:hive/hive.dart';
 import 'package:hive/src/backend/js/storage_backend_js.dart';
 import 'package:hive/src/backend/storage_backend.dart';
-import 'package:hive/src/crypto/padded_cipher.dart';
 
 class BackendManager implements BackendManagerInterface {
   @override
   Future<StorageBackend> open(
-      String name, String path, bool crashRecovery, PaddedCipher cipher) async {
+      String name, String path, bool crashRecovery, HiveCipher cipher) async {
     var db =
         await window.indexedDB.open(name, version: 1, onUpgradeNeeded: (e) {
       var db = e.target.result as Database;