+++ lib/src/backend/storage_backend_memory.dart
@@ -5,10 +5,9 @@ import 'package:hive/src/backend/storage_backend.dart';
 import 'package:hive/src/binary/frame.dart';
 import 'package:hive/src/binary/frame_helper.dart';
 import 'package:hive/src/box/keystore.dart';
-import 'package:hive/src/crypto/padded_cipher.dart';
 
 class StorageBackendMemory extends StorageBackend {
-  final PaddedCipher cipher;
+  final HiveCipher cipher;
 
   final FrameHelper frameHelper;